+++ src/repos.json
@@ -1,0 +1,3 @@
+{
+  "go-commerce-components": ["kafka-golang", "schemas", "yalometrics-golang"]
+}
